+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 2.1.0 *(2021-XX-XX)*
 * New: Add new scope: BY_HOLDER
 * Fix: Compiler error logger wasn't showing the correct message
 * New: Can annotate custom annotations with `@Injectable` to be used for injectable classes too
+* Deprecated: `Alias` is no longer useful, use environments instead
 
 Version 2.0.0 *(2020-09-05)*
 -----------------------------